Forms+Cases are being displayed on UserOpenForms/UserClosedForms. Tile counts adjusted. DateSubmitted / Date(time)Modified being displayed on the forms grids. Files modified - [Main,ProcessForms,UserOpenForms,UserClosedForms,CompanyForms,cx2,pom.xml]

+++ services/cx2/src/com/civicxpress/cx2/InspectionDesign.java
@@ -57,7 +57,7 @@ public class InspectionDesign implements Serializable {
     private String titlePrefix;
     private String prefixDate;
     private String prefixNumber;
-    private int prefixNumberStart;
+    private BigInteger prefixNumberStart;
     private int prefixNumberStep;
     private boolean prefixDashes;
     private Integer createdBy;
@@ -298,12 +298,12 @@ public class InspectionDesign implements Serializable {
         this.prefixNumber = prefixNumber;
     }
 
-    @Column(name = "`PrefixNumberStart`", nullable = false, scale = 0, precision = 10)
-    public int getPrefixNumberStart() {
+    @Column(name = "`PrefixNumberStart`", nullable = false, scale = 0, precision = 38)
+    public BigInteger getPrefixNumberStart() {
         return this.prefixNumberStart;
     }
 
-    public void setPrefixNumberStart(int prefixNumberStart) {
+    public void setPrefixNumberStart(BigInteger prefixNumberStart) {
         this.prefixNumberStart = prefixNumberStart;
     }
